shader_compiler

Function compile_slang

pub fn compile_slang(
    shader: impl AsRef<Path>,
    output_file: Option<impl AsRef<Path>>,
) -> Result<Vec<u8>>

Compiles the shader file into usable SPIRV.

This method invokes slangc in a subprocess and therefore expects slangc to be present in the system PATH.

Params

[shader] - The filesystem path to the shader’s source.

[output_file] - Optionally write the comiled shader to a file, rather than returning the bytes directly.

Returns

The compiled shader artifact as spirv words.